62 update frontmattedata class (#125)

+++ src/Actions/UpdateIndex.php
@@ -2,13 +2,11 @@
 
 namespace BenBjurstrom\Prezet\Actions;
 
-use BenBjurstrom\Prezet\Data\FrontmatterData;
-use BenBjurstrom\Prezet\Http\Controllers\ShowController;
+use BenBjurstrom\Prezet\Data\DocumentData;
 use BenBjurstrom\Prezet\Models\Document;
 use BenBjurstrom\Prezet\Models\Heading;
 use BenBjurstrom\Prezet\Models\Tag;
 use Illuminate\Support\Facades\Config;
-use Illuminate\Support\Facades\Route;
 use Illuminate\Support\Facades\Schema;
 
 class UpdateIndex
@@ -17,7 +15,7 @@ class UpdateIndex
     {
         self::ensureDatabaseExists();
 
-        $docs = app(GetAllFrontmatter::class)->handle();
+        $docs = app(GetAllDocsFromFiles::class)->handle();
 
         // Get all current slugs from filesystem
         $currentSlugs = $docs->pluck('slug')->toArray();
@@ -26,15 +24,11 @@ class UpdateIndex
         self::removeDeletedDocuments($currentSlugs);
 
         // Update or create documents
-        $docs->each(function (FrontmatterData $doc) {
+        $docs->each(function (DocumentData $doc) {
             self::upsertDocument($doc);
         });
 
         self::cleanupOrphanedTags();
-
-        Route::get('prezet/{slug}', ShowController::class)
-            ->name('prezet.show')
-            ->where('slug', '.*');
 
         UpdateSitemap::handle();
     }
@@ -48,16 +42,16 @@ class UpdateIndex
     {
         Document::whereNotIn('slug', $currentSlugs)->each(function (Document $document) {
             // This will automatically delete related headings due to cascade
-            $document->tags()->detach(); // TODO: Does this remove tags that are still in use by other documents?
+            $document->tags()->detach();
             $document->delete();
         });
     }
 
-    protected static function upsertDocument(FrontmatterData $doc): void
+    protected static function upsertDocument(DocumentData $docData): void
     {
         // Check if document exists with same slug and hash
-        $existingDoc = Document::where('slug', $doc->slug)
-            ->where('hash', $doc->hash)
+        $existingDoc = Document::where('slug', $docData->slug)
+            ->where('hash', $docData->hash)
             ->first();
 
         // If document exists with same hash, no need to update
@@ -66,38 +60,42 @@ class UpdateIndex
         }
 
         // Find document by slug to update, or create new one
-        $document = Document::where('slug', $doc->slug)->first() ?? new Document;
+        $document = Document::where('slug', $docData->slug)->first() ?? new Document;
 
-        self::updateDocumentAttributes($document, $doc);
-        self::updateHeadings($document);
+        self::updateDocumentAttributes($document, $docData);
+        self::updateHeadings($document, $docData->content);
 
-        if ($doc->tags) {
-            self::setTags($document, $doc->tags);
+        if ($docData->frontmatter->tags) {
+            self::setTags($document, $docData->frontmatter->tags);
         }
     }
 
-    protected static function updateDocumentAttributes(Document $document, FrontmatterData $doc): void
+    protected static function updateDocumentAttributes(Document $document, DocumentData $docData): void
     {
         $document->fill([
-            'slug' => $doc->slug,
-            'category' => $doc->category,
-            'draft' => $doc->draft,
-            'hash' => $doc->hash,
-            'frontmatter' => $doc,
-            'created_at' => $doc->createdAt,
-            'updated_at' => $doc->updatedAt,
+            'slug' => $docData->slug,
+            'category' => $docData->category,
+            'draft' => $docData->draft,
+            'hash' => $docData->hash,
+            'frontmatter' => $docData->frontmatter,
+            'created_at' => $docData->createdAt,
+            'updated_at' => $docData->updatedAt,
         ]);
 
         $document->save();
     }
 
-    protected static function updateHeadings(Document $document): void
+    protected static function updateHeadings(Document $document, ?string $content): void
     {
+        if (! $content) {
+            // This shouldn't happen, but just in case
+            throw new \RuntimeException('Cannot create headings. Document content is empty.');
+        }
+
         // Delete existing headings
         $document->headings()->delete();
 
-        $md = GetMarkdown::handle($document->filepath);
-        $html = ParseMarkdown::handle($md);
+        $html = ParseMarkdown::handle($content);
         $headings = GetFlatHeadings::handle($html);
         $title = $document->frontmatter->title;
